+++ Content/Items/Accessories/Mayday/RedAlert.cs
@@ -19,16 +19,9 @@ namespace SuperfluityTwo.Content.Items.Accessories.Mayday
         public override void AddRecipes()
         {
             CreateRecipe()
-                .AddIngredient(ItemID.DemoniteBar, 12)
+                .AddIngredient(ItemID.HellstoneBar, 12)
                 .AddIngredient(ItemID.Ruby, 5)
                 .AddTile(TileID.Anvils)
-                .AddDecraftCondition(Condition.CorruptWorld)
-                .Register();
-            CreateRecipe()
-                .AddIngredient(ItemID.CrimtaneBar, 12)
-                .AddIngredient(ItemID.Ruby, 5)
-                .AddTile(TileID.Anvils)
-                .AddDecraftCondition(Condition.CrimsonWorld)
                 .Register();
         }